# image_service/api/internal_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Internal HTTP API сервиса изображений.",
        "title": "Image Service Internal API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/admin/image/{id}": {
            "get": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "Возвращает бинарное содержимое изображения без гостевой обработки",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Получение изображения (администратор)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID изображения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Тип изображения (icon)",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/v1.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.APIResponse"
                        }
                    }
                }
            }
        },
        "/guest/image/{id}": {
            "get": {
                "description": "Возвращает бинарное содержимое изображения для неавторизованных пользователей",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Получение изображения (гостевой доступ)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID изображения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Тип изображения (icon)",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.APIResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Проверяет доступность сервиса",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.APIResponse"
                        }
                    }
                }
            }
        },
        "/image": {
            "get": {
                "description": "Возвращает все изображения в сервисе",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Список всех изображений с пагинацией",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Номер страницы (с нуля)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Размер страницы",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/v1.UnmoderatedImagesResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.APIResponse"
                        }
                    }
                }
            }
        },
        "/image/meta/{id}": {
            "get": {
                "description": "Возвращает метаинформацию об изображении без бинарного содержимого",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Метаданные изображения",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID изображения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/v1.ImageMetaResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.APIResponse"
                        }
                    }
                }
            }
        },
        "/image/unmoderated": {
            "get": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "Возвращает страницу изображений со статусом unmoderated",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Список немодерированных изображений",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Номер страницы (с нуля)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Размер страницы",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/v1.UnmoderatedImagesResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/v1.APIResponse"
                        }
                    }
                }
            }
        },
        "/image/upload": {
            "post": {
                "description": "Сохраняет новое изображение со статусом unmoderated",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Загрузка изображения",
                "parameters": [
                    {
                        "description": "Данные изображения",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.SwaggerAddImageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/v1.ImageMetaResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.APIResponse"
                        }
                    }
                }
            }
        },
        "/image/{id}/approve": {
            "put": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "Устанавливает изображению статус approved",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Одобрение изображения",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID изображения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/v1.ImageStatusResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/v1.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.APIResponse"
                        }
                    }
                }
            }
        },
        "/image/{id}/block": {
            "put": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "Устанавливает изображению статус blocked",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Блокировка изображения",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID изображения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/v1.ImageStatusResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/v1.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "v1.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "err_message": {
                    "type": "string",
                    "example": ""
                },
                "spread_id": {
                    "type": "string",
                    "example": "uuid-запроса"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "v1.ImageListItem": {
            "type": "object",
            "properties": {
                "crated_at": {
                    "type": "string",
                    "example": "2026-07-03T09:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "media_type": {
                    "type": "string",
                    "example": "jpeg"
                },
                "name": {
                    "type": "string",
                    "example": "test_image"
                },
                "status": {
                    "type": "string",
                    "example": "unmoderated"
                }
            }
        },
        "v1.ImageMetaResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-07-03T09:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "media_type": {
                    "type": "string",
                    "example": "jpeg"
                },
                "name": {
                    "type": "string",
                    "example": "test_image"
                },
                "status": {
                    "type": "string",
                    "example": "unmoderated"
                }
            }
        },
        "v1.ImageStatusResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "test_image"
                },
                "status": {
                    "type": "string",
                    "example": "blocked"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-07-03T09:00:00Z"
                }
            }
        },
        "v1.SwaggerAddImageRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "\u003cbase64\u003e"
                },
                "media_type": {
                    "type": "string",
                    "example": "jpeg"
                },
                "name": {
                    "type": "string",
                    "example": "test_image"
                }
            }
        },
        "v1.UnmoderatedImagesResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 1
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.ImageListItem"
                    }
                },
                "total_count": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    },
    "securityDefinitions": {
        "AccessToken": {
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}